[Modify] vuex 모듈화 작업
