+++ happy-house/src/components/qna/QnaComment.vue
@@ -25,7 +25,6 @@
 </template>
 
 <script type="module">
-// import http from "@/api/index.js";
 import { modifyQnaComment, deleteQnaComment } from "@/api/qna";
 
 export default {
@@ -67,13 +66,6 @@ export default {
             console.log("QnA 댓글 수정 오류 : " + error);
           }
         );
-        // http.put(`/qnacomment/${this.comment.id}`, this.changedComment).then(({ data }) => {
-        //   if (data.flag === "success") {
-        //     alert("답글을 수정하였습니다.");
-        //     this.$emit("changData", this.idx, this.changedComment);
-        //     this.isModifyStatus = false;
-        //   }
-        // });
       }
     },
     deleteComment() {
@@ -90,12 +82,6 @@ export default {
             console.log("QnA 댓글 삭제 오류 : " + error);
           }
         );
-        // http.delete(`/qnacomment/${this.comment.id}`).then(({ data }) => {
-        //   if (data.flag === "success") {
-        //     alert("답글을 삭제하였습니다.");
-        //     this.$router.go();
-        //   }
-        // });
       }
     },
   },
