+++ happy-house/src/components/qna/QnaModify.vue
@@ -46,7 +46,6 @@
 </template>
 
 <script>
-// import http from "@/api";
 import { getQnaDetail, modifyQna } from "@/api/qna";
 
 export default {
@@ -75,15 +74,6 @@ export default {
         console.log("QnaModify의 QnA 상세보기 오류 : " + error);
       }
     );
-    // http.get(`/qna/${this.$route.params.id}`).then(({ data }) => {
-    //   if (data.flag === "success") {
-    //     this.question = data.data[0];
-    //     this.origin.title = this.question.title;
-    //     this.origin.content = this.question.content;
-    //   } else {
-    //     // 추후
-    //   }
-    // });
   },
   methods: {
     modifyQna() {
@@ -105,14 +95,6 @@ export default {
         },
         (error) => console.log("QnA 수정 오류 : " + error)
       );
-      // http.put(`/qna/${this.question.id}`, this.sendingData).then(({ data }) => {
-      //   if (data.flag === "success") {
-      //     alert("글 수정 완료!!");
-      //     this.$router.push({ name: "qnadetail", params: { id: this.question.id } });
-      //   } else {
-      //     // 추후
-      //   }
-      // });
     },
     cancelModifyBtn() {
       this.$router.push({ name: "qnadetail", params: { id: this.question.id } });
