+++ happy-house/src/store/index.js
@@ -1,18 +1,9 @@
 import Vue from "vue";
 import Vuex from "vuex";
+import userStore from "@/store/modules/userStore";
 
 Vue.use(Vuex);
 
 export default new Vuex.Store({
-  state: {
-    isLoginStatus: false,
-  },
-  getters: {},
-  mutations: {
-    SET_IS_LOGIN_STATUS(state, isLoginStatus) {
-      state.isLoginStatus = isLoginStatus;
-    },
-  },
-  actions: {},
-  modules: {},
+  modules: { userStore },
 });
